+++ macaw-aarch32-symbolic/tests/pass/so/getone.c
@@ -1,0 +1,5 @@
+#include "getone.h"
+
+int getone(void) {
+  return 1;
+}