+++ x86/src/Data/Macaw/X86.hs
@@ -26,6 +26,7 @@ module Data.Macaw.X86
        , x86_64_freeBSD_info
        , x86_64_linux_info
        , x86_64CallParams
+       , x86_64PLTStubInfo
        , freeBSD_syscallPersonality
        , linux_syscallPersonality
          -- * Low level exports
@@ -58,6 +59,7 @@ import           Control.Lens
 import           Control.Monad.Cont
 import           Control.Monad.Except
 import           Control.Monad.ST
+import qualified Data.ElfEdit as EE
 import           Data.Foldable
 import qualified Data.Map as Map
 import           Data.Parameterized.Classes
@@ -81,6 +83,7 @@ import           Data.Macaw.CFG
 import           Data.Macaw.CFG.DemandSet
 import           Data.Macaw.Discovery ( defaultClassifier )
 import qualified Data.Macaw.Memory as MM
+import qualified Data.Macaw.Memory.ElfLoader.PLTStubs as MMEP
 import qualified Data.Macaw.Memory.Permissions as Perm
 import           Data.Macaw.Types
   ( n8
@@ -615,6 +618,13 @@ x86_64_linux_info :: ArchitectureInfo X86_64
 x86_64_linux_info = x86_64_info preserveFn
   where preserveFn r = Set.member (Some r) linuxSystemCallPreservedRegisters
 
+-- | PLT stub information for X86_64 relocation types.
+x86_64PLTStubInfo :: MMEP.PLTStubInfo EE.X86_64_RelocationType
+x86_64PLTStubInfo = MMEP.PLTStubInfo
+  { MMEP.pltFunSize     = 16
+  , MMEP.pltStubSize    = 16
+  , MMEP.pltGotStubSize = 8
+  }
 
 {-# DEPRECATED disassembleBlock "Use disassembleFn x86_64_info" #-}