+++ x86_symbolic/tests/Main.hs
@@ -75,8 +75,8 @@ main = do
   -- These are pass/fail in that the assertions in the "pass" set are true (and
   -- the solver returns Unsat), while the assertions in the "fail" set are false
   -- (and the solver returns Sat).
-  passTestFilePaths <- SFG.glob "tests/pass/*.exe"
-  failTestFilePaths <- SFG.glob "tests/fail/*.exe"
+  passTestFilePaths <- SFG.glob "tests/pass/**/*.exe"
+  failTestFilePaths <- SFG.glob "tests/fail/**/*.exe"
   let passRes = MST.SimulationResult MST.Unsat
   let failRes = MST.SimulationResult MST.Sat
   let passTests = TT.testGroup "True assertions" (map (mkSymExTest passRes) passTestFilePaths)
@@ -110,8 +110,8 @@ mkSymExTest expected exePath = TT.askOption $ \saveSMT@(SaveSMT _) -> TT.askOpti
       case Elf.headerClass (Elf.header ehi) of
         Elf.ELFCLASS32 -> TTH.assertFailure "32 bit x86 binaries are not supported"
         Elf.ELFCLASS64 -> do
-          (mem, discState) <- MST.runDiscovery ehi MST.toAddrSymMap MX.x86_64_linux_info
-          let funInfos = Map.elems (discState ^. M.funInfo)
+          binariesInfo <- MST.runDiscovery ehi exePath MST.toAddrSymMap MX.x86_64_linux_info MX.x86_64PLTStubInfo
+          let funInfos = Map.elems (MST.binaryDiscState (MST.mainBinaryInfo binariesInfo) ^. M.funInfo)
           let testEntryPoints = mapMaybe hasTestPrefix funInfos
           F.forM_ testEntryPoints $ \(name, Some dfi) -> do
             step ("Testing " ++ BS8.unpack name)
@@ -132,7 +132,7 @@ mkSymExTest expected exePath = TT.askOption $ \saveSMT@(SaveSMT _) -> TT.askOpti
               let extract = x86ResultExtractor archVals
               logger <- makeGoalLogger saveSMT solver name exePath
               let ?memOpts = LLVM.defaultMemOptions
-              simRes <- MST.simulateAndVerify solver logger bak execFeatures MX.x86_64_linux_info archVals mem extract discState dfi
+              simRes <- MST.simulateAndVerify solver logger bak execFeatures MX.x86_64_linux_info archVals binariesInfo extract dfi
               TTH.assertEqual "AssertionResult" expected simRes
 
 writeMacawIR :: (MC.ArchConstraints arch) => SaveMacaw -> String -> M.DiscoveryFunInfo arch ids -> IO ()